Added support for pushing locals as pointers to values on the stack. Handle STOPOP and MUL

+++ m8rscript/Scanner.cpp
@@ -38,6 +38,7 @@ POSSIBILITY OF SUCH DAMAGE.
 #include "Parser.h"
 
 #include <cstring>
+#include <string>
 
 using namespace m8r;
 
@@ -433,7 +434,7 @@ uint8_t Scanner::getToken(YYSTYPE* tokenValue)
                     if (token == T_INTEGER) {
                         tokenValue->integer = static_cast<uint32_t>(strtol(_tokenString.c_str(), NULL, 0));
                     } else {
-                        tokenValue->number = static_cast<float>(atof(_tokenString.c_str()));
+                        tokenValue->number = static_cast<float>(std::stof(_tokenString.c_str()));
                     }
                     _tokenString.erase();
 					break;
